LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/webvtt-cross-origin.https-expected.txt


PASS initialize global state 
PASS same-origin text track should load 
FAIL cross-origin text track with no-cors request should not load assert_equals: expected "error event" but got "load event"
PASS cross-origin text track with rejected cors request should not load 
FAIL cross-origin text track with approved cors request should not load assert_equals: expected "error event" but got "load event"
PASS same-origin text track that redirects same-origin should load 
FAIL same-origin text track that redirects cross-origin should not load assert_equals: expected "error event" but got "load event"
PASS same-origin text track that redirects to a cross-origin text track with rejected cors should not load 
FAIL same-origin text track that redirects to a cross-origin text track with approved cors should not load assert_equals: expected "error event" but got "load event"
PASS restore global state 
